+++ R/Htest.R
@@ -17,8 +17,9 @@
 #'
 #' @importFrom stats formula complete.cases lm nobs pchisq
 #' @export
-#' @references Hoffman, L. (2007). Multilevel models for examining individual differences in within-person variation and covariation over time. \emph{Multivariate Behavioral Research, 42}(4), 609–629.
-#' @references Raudenbush, S., & Bryk, A. (2002). \emph{Hierarchical linear models: Applications and data analysis methods (2nd ed.)}. Sage.
+#' @references
+#' \cite{Hoffman, L. (2007). Multilevel models for examining individual differences in within-person variation and covariation over time. \emph{Multivariate Behavioral Research, 42}(4), 609–629.}
+#' \cite{Raudenbush, S., & Bryk, A. (2002). \emph{Hierarchical linear models: Applications and data analysis methods (2nd ed.)}. Sage.}
 #' @examples
 #' set.seed(123)
 #' x1 <- rnorm(400)