+++ NEWS.md
@@ -1,3 +1,7 @@
+# MLMusingR 0.4.0
+* Added mixPV, lrtPV to allow for models using weights with the WeMix package.
+* Added the wscale function to scale weights ala Mplus.
+
 # MLMusingR 0.2.0
 * Revised robust_mixed (w/Bixi Zhang).
 * Improved printing output.
@@ -5,6 +9,5 @@
 * Modified wording in nmiss.
 
 # MLMusingR 0.1.0
-
 * Added a `NEWS.md` file to track changes to the package.
 * Initial release.